requirements.txt
httpx
python-dotenv
jinja2
